Add Julia implementation of counting
count.jl reads a JSON description of a DFA from stdin, calculates the
number of strings in the language accepted by the DFA for a range of
string lengths, and prints each count on a separate line to stdout.

The range of lengths is given by the program's arguments:

1 arg: [arg1,arg1]
2 args: [arg1,arg2)
otherwise: [0,infinity)

In the last mode, the program will exit gracefully when stdout is closed.

An example invocation:

```
./count.jl < exampleDFA.json | head -100 | tail -1
```